substantial progress
The web server basics are essentially done, with lots of optimizations
including gzip compression completed. The database system is mostly
working, with tables applying to databases correctly and replicably. The
next step is to create a good system for redeploying the entities -
current idea is to make a temporary SQLite file, deploy the database
into that, generate the entity files, and then remove the sqlite file;
as such entity files likely shouldn't be included in the git repository.
This is incredible progress however and leaves the toolchain set up for
the deployment of Flowbite (Tailwind for Svelte).

Signed-off-by: Amy Parker <amy@amyip.net>